block: fiops sync/async scale
CFQ gives 2.5 times more share to sync workload. This matches CFQ.

Note this is different with the read/write scale. We have 3 types of
requests:
1. read
2. sync write
3. write
CFQ doesn't differentitate type 1 and 2, but request cost of 1 and 2
are usually different for flash based storage. So we have both sync/async
and read/write scale here.

Change-Id: I10e0405dce90415777d1196fa90f146fc937dbe0
Signed-off-by: Shaohua Li <shaohua.li@intel.com>
Signed-off-by: RobbieL811 <RobbieLancaster811@gmail.com>